kauth: Use libtool
Modify the kauth directory so that we use libtool to build its library.
Unlike most of our other directories, kauth has few pthreaded consumers
within the tree, so the libtool library itself is currently unused. The
libtool conversion is done as a stepping point for libafsauthent, which
does include kauth objects.

At the same time, we take the opportunity to modify kaserver so that it
uses libkauth, rather than referencing the library objects directly

Change-Id: I9d261686753b30c415f767603d44c7ce6e10eed1
Reviewed-on: http://gerrit.openafs.org/8095
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Derrick Brashear <shadow@your-file-system.com>