libafs*: Don't re-export opr/util functions
On Solaris, a library can't re-export functions which come from one
of its dependencies - so libafsrpc can't easily republish the 5 opr
functions that are currently in the libafsrpc map file.
Including the bits of the opr library that define these functions
causes problems with then linking libafsrpc against libopr, which
we do in a number of places.

libafsauthent has exactly the same problem, with a few symbols that
it pulls in from util.

The affected symbols were all added relatively recently, and have
limited uses outside of the OpenAFS tree, so just remove them from
the export list so we are one step closer to building on Solaris
again.

Change-Id: If0a1105320e416ebc013ff670dddd5d247b06f6b
Reviewed-on: http://gerrit.openafs.org/8102
Reviewed-by: Derrick Brashear <shadow@your-file-system.com>
Tested-by: Derrick Brashear <shadow@your-file-system.com>