rxkad: silence unused variable warnings in v5gen.c
Sprinkle AFS_UNUSED attributes where the compiler complains
about unused variables.

Note that this file is assembled from code generated by the heimdal
source.  A similar fix exists in the generator code upstream, so
when the file is re-generated in the future it should not be
necessary to re-apply this.

Change-Id: I2ef407428d90157050481d5630f6f22510d65ca3
Reviewed-on: http://gerrit.openafs.org/8122
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Derrick Brashear <shadow@your-file-system.com>
Reviewed-by: Jeffrey Altman <jaltman@your-file-system.com>